+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Schachtabelle Changelog
 
+## Version 3.0.2 (2022-03-01)
+
+* Fix: IRN-Flagge wird nicht erkannt (Flagge zusätzlich in Alpha3-Code suchen, wenn in IOC-Code nicht gefunden)
+* Add: Flaggen-Array um Nordirland, Wales, England und Schottland erweitert
+
 ## Version 3.0.1 (2022-02-24)
 
 * Fix: Warning: Invalid argument supplied for foreach() - wenn tl_content.chesstable_markierungen leer ist